+++ R/get_code.R
@@ -6,6 +6,8 @@
 #' @param name character. Optional argument giving the name to assign to the object.
 #' @param sub logical. Should states or models be referenced by name in \code{define_model}
 #'  and \code{run_model} instead of including the entire code?
+#' @param depth Depth of the function call.
+#' @param n_space Number of space used for indentation.
 #' @param ... Additional parameters passed to methods.
 #'
 #' @return A character string.
@@ -50,27 +52,37 @@ to_text_dots <- function(x, name = TRUE) {
   }
 }
 
+left_pad <- function(x, n_pad, prefix = FALSE) {
+  pad <- paste(rep(" ", n_pad), collapse = "")
+  pre <- if (prefix) pad else ""
+  paste0(pre, gsub("\\n", paste0("\\\n", pad), x))
+}
+
 #' @export
 #' @rdname get_code
-get_code.uneval_parameters <- function(x, name = NULL, ...) {
+get_code.uneval_parameters <- function(x, name = NULL,
+                                       depth = 0, n_space = 2,
+                                       ...) {
   if (! is.null(name)) {
     name <- paste(name, "<- ")
   }
   structure(
-    paste0(
+    left_pad(paste0(
       name,
       "define_parameters(\n  ",
       paste(to_text_dots(x), collapse = ",\n  "),
       "\n)",
       collapse = ""
-    ),
+    ), n_pad = depth * n_space),
     class = "heemod_code"
   )
 }
 
 #' @export
 #' @rdname get_code
-get_code.uneval_matrix <- function(x, name = NULL, ...) {
+get_code.uneval_matrix <- function(x, name = NULL,
+                                   depth = 0, n_space = 2,
+                                   ...) {
   if (! is.null(name)) {
     name <- paste(name, "<- ")
   }
@@ -79,7 +91,7 @@ get_code.uneval_matrix <- function(x, name = NULL, ...) {
   sn <- get_state_names(x)
   
   structure(
-    paste0(
+    left_pad(paste0(
       name,
       "define_matrix(\n  ",
       "state_names = ",
@@ -90,47 +102,54 @@ get_code.uneval_matrix <- function(x, name = NULL, ...) {
       ),
       "\n)",
       collapse = ""
-    ),
+    ), n_pad = depth * n_space),
     class = "heemod_code"
   )
 }
 
 #' @export
 #' @rdname get_code
-get_code.state <- function(x, name = NULL, ...) {
+get_code.state <- function(x, name = NULL,
+                           depth = 0, n_space = 2,
+                           ...) {
   if (! is.null(name)) {
     name <- paste(name, "<- ")
   }
   
   structure(
-    paste0(
+    left_pad(paste0(
       name,
       "define_state(\n  ",
       paste(to_text_dots(x), collapse = ",\n  "),
       "\n)",
       collapse = ""
-    ),
+    ), n_pad = 0),
     class = "heemod_code"
   )
 }
 
 #' @export
 #' @rdname get_code
-get_code.uneval_state_list <- function(x, ...) {
+get_code.uneval_state_list <- function(x,
+                                       depth = 0, n_space = 2,
+                                       ...) {
+  
   structure(
-    paste(
+    left_pad(paste(
       names(x),
-      unlist(lapply(x, get_code)),
+      unlist(lapply(x, get_code, depth = depth)),
       sep = " = ",
-      collapse = ",\n  "
-    ),
+      collapse = ",\n"
+    ), n_pad = depth * n_space, prefix = TRUE),
     class = "heemod_code"
   )
 }
 
 #' @export
 #' @rdname get_code
-get_code.uneval_model <- function(x, name = NULL, sub = FALSE, ...) {
+get_code.uneval_model <- function(x, name = NULL, sub = FALSE,
+                                  depth = 0, n_space = 2,
+                                  ...) {
   if (! is.null(name)) {
     name <- paste(name, "<- ")
   }
@@ -143,26 +162,28 @@ get_code.uneval_model <- function(x, name = NULL, sub = FALSE, ...) {
       collapse = ",\n  "
     )
   } else {
-    st <- get_code(x$states)
+    st <- get_code(x$states, depth = depth)
   }
   
   structure(
-    paste0(
+    left_pad(paste0(
       name,
       "define_model(\n  ",
       "transition_matrix = ",
-      get_code(x$transition_matrix),
-      ",\n  ",
+      get_code(x$transition_matrix, depth = depth),
+      ",\n",
       st,
       "\n)"
-    ),
+    ), n_pad = depth * n_space),
     class = "heemod_code"
   )
 }
 
 #' @export
 #' @rdname get_code
-get_code.eval_model_list <- function(x, name = NULL, sub = FALSE, ...) {
+get_code.run_models <- function(x, name = NULL, sub = FALSE,
+                                depth = 0, n_space = 2,
+                                ...) {
   if (! is.null(name)) {
     name <- paste(name, "<- ")
   }
@@ -172,25 +193,27 @@ get_code.eval_model_list <- function(x, name = NULL, sub = FALSE, ...) {
       names(attr(x, "uneval_model_list")),
       make_names(paste0("m_", names(attr(x, "uneval_model_list")))),
       sep = " = ",
-      collapse = ",\n  "
+      collapse = ",\n"
     )
   } else {
+    
     md <- paste(
       names(attr(x, "uneval_model_list")),
-      unlist(lapply(attr(x, "uneval_model_list"), get_code)),
+      unlist(lapply(attr(x, "uneval_model_list"),
+                    get_code, depth = depth + 1)),
       sep = " = ",
       collapse = ",\n  "
     )
   }
   
   structure(
-    paste0(
+    left_pad(paste0(
       name,
       "run_models(\n  ",
       md,
       ",\n  ",
       "parameters = ",
-      get_code(attr(x, "parameters")),
+      get_code(attr(x, "parameters"), depth = depth + 1),
       ",\n  ",
       "init = ",
       paste0("c(", paste(attr(x, "init"), collapse = ", "), "),\n  ", collapse = ""),
@@ -209,7 +232,7 @@ get_code.eval_model_list <- function(x, name = NULL, sub = FALSE, ...) {
       "effect = ",
       deparse(attr(x, "ce")[[2]]$expr, width.cutoff = 500L),
       "\n)"
-    ),
+    ), n_pad = depth * n_space),
     class = "heemod_code"
   )