+++ R/run_model_summary.R
@@ -1,5 +1,5 @@
 #' @export
-print.eval_model_list <- function(x, ...) {
+print.run_models <- function(x, ...) {
   print(summary(x, ...))
 }
 #' Summarise Markov Model Results
@@ -10,41 +10,43 @@ print.eval_model_list <- function(x, ...) {
 #'   
 #' @return A \code{summary_eval_model_list} object.
 #' @export
-#' 
-summary.eval_model_list <- function(object, ...) {
+summary.run_models <- function(object, ...) {
   if (! all(c(".cost", ".effect") %in% names(object))) {
     warning("No cost and/or effect defined, model summary unavailable.")
     return(invisible(NULL))
   }
   
   res <- as.data.frame(compute_icer(normalize_ce(object)))
+  rownames(res) <- res$.model_names
+  res <- dplyr::select_(res, quote(- .model_names))
   
-  res <- dplyr::select(res, - .model_names)
   
-  rownames(res) <- object$.model_names
-  
-  res_comp <- res[c(".cost", ".effect", ".icer")]
+  res_comp <- res[c(".dcost", ".deffect", ".icer")]
   is.na(res_comp$.icer) <- ! is.finite(res_comp$.icer)
   res_comp$.icer <- format(res_comp$.icer)
   res_comp$.icer[res_comp$.icer == "NA"] <- "-"
-  res_comp$.cost <- res_comp$.cost / sum(attr(object, "init"))
-  res_comp$.effect <- res_comp$.effect / sum(attr(object, "init"))
-  names(res_comp) <- c("Cost", "Effect", "ICER")
+  res_comp$.dcost <- res_comp$.dcost / sum(attr(object, "init"))
+  res_comp$.deffect <- res_comp$.deffect / sum(attr(object, "init"))
   
   structure(
     list(
-      res = dplyr::select(res, - .cost, - .effect, - .icer),
-      res_comp = res_comp[-1, ],
+      res = dplyr::select_(
+        res,
+        quote(- .cost),
+        quote(- .effect),
+        quote(- .icer),
+        quote(- .dcost),
+        quote(- .deffect)
+      ),
+      res_comp = res_comp,
       cycles = attr(object, "cycles"),
       init = attr(object, "init"),
       count_args = attr(object, "count_args"),
       frontier = get_frontier(object)
     ),
-    class = "summary_eval_model_list"
+    class = "summary_run_models"
   )
 }
-if(getRversion() >= "2.15.1")
-  utils::globalVariables(c(".model_names", ".cost", ".effect", ".icer"))
 
 #' Normalize Cost and Effect
 #' 
@@ -59,7 +61,7 @@ if(getRversion() >= "2.15.1")
 normalize_ce <- function(x) {
   UseMethod("normalize_ce")
 }
-normalize_ce.eval_model_list <- function(x) {
+normalize_ce.run_models <- function(x) {
   bm <- get_base_model(x)
   x$.cost <- x$.cost - x$.cost[x$.model_names == bm]
   x$.effect <- x$.effect - x$.effect[x$.model_names == bm]
@@ -73,27 +75,33 @@ normalize_ce.eval_model_list <- function(x) {
 #' Models are ordered by effectiveness and ICER are computed sequencially.
 #' 
 #' @param x Result of \code{\link{run_models}}.
+#' @param model_order Order in which the models 
+#' should be sorted. Default: by increasing effect.
 #'   
 #' @return A \code{data.frame} with computed ICER.
 #' @export
 #' 
-compute_icer <- function(x) {
-  tab <- x[order(x$.effect), ]
+compute_icer <- function(x, model_order = order(x$.effect)) {
+  
+  tab <- x[model_order, ]
   
   tab$.icer <- NA
+  tab$.dcost <- NA
+  tab$.deffect <- NA
   for (i in seq_len(nrow(tab))) {
     if (i == 1) {
       tab$.icer[i] <- -Inf
     } else {
-      tab$.icer[i] <- (tab$.cost[i] - tab$.cost[i-1]) /
-        (tab$.effect[i] - tab$.effect[i-1])
+      tab$.dcost[i] <- tab$.cost[i] - tab$.cost[i-1]
+      tab$.deffect[i] <- tab$.effect[i] - tab$.effect[i-1]
+      tab$.icer[i] <- tab$.dcost[i] / tab$.deffect[i]
     }
   }
   tab
 }
 
 #' @export
-print.summary_eval_model_list <- function(x, ...) {
+print.summary_run_models <- function(x, ...) {
   cat(sprintf(
     "%i Markov model%s run for %i cycle%s.\n\n",
     nrow(x$res),
@@ -115,6 +123,6 @@ print.summary_eval_model_list <- function(x, ...) {
     cat("\nEfficiency frontier:\n\n")
     cat(x$frontier)
     cat("\n\nModel difference:\n\n")
-    print(x$res_comp)
+    print(setNames(x$res_comp[-1, ], c("Cost", "Effect", "ICER")))
   }
 }