+++ tests/testthat/test_model.R
@@ -50,6 +50,14 @@ test_that(
     )
     expect_error(
       define_model(
+        transition_matrix = mat2,
+        s1,
+        s2,
+        s2
+      )
+    )
+    expect_error(
+      define_model(
         transition_matrix = mat1,
         X1 = s1,
         X3 = s2
@@ -218,8 +226,8 @@ Initial states:
 X1 1
 X2 0
           x        y
-I  1170.000  615.000
-II 1592.538 1514.507
+II 1170.000  615.000
+I  1592.538 1514.507
 
 Efficiency frontier:
 
@@ -227,8 +235,8 @@ II I
 
 Model difference:
 
-       Cost   Effect      ICER
-II 422.5384 899.5074 0.4697442",
+      Cost   Effect      ICER
+I 422.5384 899.5074 0.4697442",
       fixed = TRUE
     )
     expect_output(
@@ -241,7 +249,7 @@ II 422.5384 899.5074 0.4697442",
     )
     expect_output(
       print(summary(e_mod2)),
-      '2 Markov models run for 5 cycles.
+      "2 Markov models run for 5 cycles.
 
 Initial states:
 
@@ -249,8 +257,8 @@ Initial states:
 X1 1
 X2 0
           x        y
-I  1170.000  615.000
-II 1592.538 1514.507
+II 1170.000  615.000
+I  1592.538 1514.507
 
 Efficiency frontier:
 
@@ -258,8 +266,8 @@ II I
 
 Model difference:
 
-       Cost   Effect      ICER
-II 422.5384 899.5074 0.4697442',
+      Cost   Effect      ICER
+I 422.5384 899.5074 0.4697442",
       fixed = TRUE
     )
     expect_output(
@@ -281,8 +289,8 @@ Initial states:
 X1 1
 X2 0
             x        y
-mod1 1170.000  615.000
-mod2 1592.538 1514.507
+mod2 1170.000  615.000
+mod1 1592.538 1514.507
 
 Efficiency frontier:
 
@@ -291,7 +299,7 @@ mod2 mod1
 Model difference:
 
          Cost   Effect      ICER
-mod2 422.5384 899.5074 0.4697442",
+mod1 422.5384 899.5074 0.4697442",
       fixed = TRUE
     )
   }